refactor(bin-node): refactor the code of bin node 1. add `paths` to `kuai-common` package 2. move the download of binary file to the class
