+++ packages/common/package.json
@@ -8,6 +8,7 @@
     "test": "jest"
   },
   "dependencies": {
-    "@ckb-lumos/lumos": "0.19.0"
+    "@ckb-lumos/lumos": "0.19.0",
+    "undici": "5.22.1"
   }
 }
