+++ packages/common/src/constant.ts
@@ -1,0 +1,3 @@
+import envPaths from 'env-paths'
+
+export const PATH = envPaths('kuai')
